Moved reverse utility functions for consistency
